refactor: extract clip path styles
