+++ components/Socials.tsx
@@ -1,6 +1,9 @@
+import Link from "./ui/link"
+
 import { cn } from "@/lib/utils"
 
-import Link from "./ui/link"
+import * as clipPaths from "@/styles/clipPaths"
+
 import GitHub from "./svgs/github.svg"
 import Mail from "./svgs/mail.svg"
 import Twitter from "./svgs/twitter.svg"
@@ -16,10 +19,7 @@ const Socials = () => (
         "transition-transform hover:rotate-2 hover:bg-body hover:text-accent-dark hover:transition-transform"
       )}
       hideArrow
-      style={{
-        clipPath:
-          "polygon(0 0, 100% 0, 100% 100%, 0.75em 100%, 0 calc(100% - 0.75em))",
-      }}
+      style={clipPaths.mediumBottomLeft}
     >
       <Mail className="scale-100 transition-transform group-hover:scale-125 group-hover:transition-transform" />
     </Link>
@@ -44,10 +44,7 @@ const Socials = () => (
         "transition-transform hover:rotate-2 hover:bg-body hover:text-accent-alt hover:transition-transform"
       )}
       hideArrow
-      style={{
-        clipPath:
-          "polygon(0 0, calc(100% - 0.75em) 0, 100% 0.75em, 100% 100%, 0 100%)",
-      }}
+      style={clipPaths.mediumUpperRight}
     >
       <GitHub className="scale-100 transition-transform group-hover:scale-125 group-hover:transition-transform" />
     </Link>
